refactor: update ui mappers and add frontend tests
